refactor: move hooks to props of the nluTable
