+++ botfront/imports/ui/components/nlu/models/NluTable.jsx
@@ -3,24 +3,23 @@ import DataTable from '../../common/DataTable';
 import IntentLabel from '../common/IntentLabel';
 import IconButton from '../../common/IconButton';
 import { clearTypenameField } from '../../../../lib/client.safe.utils';
-import {
-    useExamples, useDeleteExamples, useUpdateExample, useSwitchCannonical,
-} from './hooks';
+
 import { wrapMeteorCallback } from '../../utils/Errors';
 import { _appendSynonymsToText } from '../../../../lib/filterExamples';
 
 
 function NluTable(props) {
-    const { projectId, workingLanguage, entitySynonyms } = props;
-
-
-    const variables = { projectId, language: workingLanguage, pageSize: 20 };
     const {
-        data, loading: loadingExamples, hasNextPage, loadMore,
-    } = useExamples(variables);
-    const [deleteExamples] = useDeleteExamples(variables);
-    const [switchCanonical] = useSwitchCannonical(variables);
-    const [updateExample] = useUpdateExample(variables);
+        entitySynonyms,
+        loadingExamples,
+        data,
+        updateExample,
+        deleteExamples,
+        switchCanonical,
+        loadMore,
+        hasNextPage,
+    } = props;
+
     const tableRef = useRef(null);
     const [examples, setExamples] = useState([]);
     const [selection, setSelection] = useState([]);
